Integrated -quiet in the same way as other clag flags
Disabled output reformatting as it interferes with file navigation and error logging in the Visual Studio extension.
The purpose of reformatting was to fix the file navigation is VS Extension. A fix will be provided in the VS extension.

Removed the prefix "Error:" when an error is encountered. It intereferes with parsing in the VS Extension.
The foreground is changed to red so in the console the user sees it's an error.