Support for custom Clang "bin" directories
If the Clang binaries are not in %PATH% and if Clang cannot be installed in "Program Files" due to missing rights, the user can manually specify the location of the Clang "bin" directory in the "General" section of the plugin settings page.

When specified, this path is transmitted to the PowerShell script via the "-clang-bin-dir" parameter.

The script will search for the Clang binaries in the following order:
- %PATH%
- the manual Clang "bin" directory override (the "-clang-bin-dir" parameter, if present)
- %ProgramW6432%\LLVM\bin (64-bit Clang default installed on 64-bit Windows)
- %ProgramFiles(x86)%\LLVM\bin (32-bit Clang default installed on 64-bit Windows)