Only compile source files if they do not have the ExcludedFromBuild child property set to true.
This fixes an issue where clang++ or clang-tidy fail on source files that
are not compatible with the active build configuration or platform,
because they have been conditionally excluded in the project file.

The conditional checks that filter source files in Get-ProjectFilesToCompile
have been pulled out into a new Should-CompileFile function, where an
additional check for the value of an optional ExcludedFromBuild child
property has been added.