Correctly handle empty project files.
This is a fix for an exception being thrown from Get-ProjectFilesToCompile when attempting to run Clang Power Tools on an empty .vcxproj file.

When a project is empty, the file nodes are null, which were trying to be bound to the mandatory fileNode parameter of Should-CompileFile. This error was likely happening before the prior change to support the ExcludedFromBuild MSBuild property, as it was trying to dereference the Include field of each node.

In our particular case, some of our solutions have an empty project in them for certain continuous integration scripts to use on our build servers, so this fix addresses the problem.