Fix #1230
XmlElementHasUnsatisfiedCondition is effectively the same code that was inside SanitizeProjectNode's `"ItemGroup"` handler which was already evaluating conditions in its child nodes. This new helper is now used for ItemGroup, but also the handler for ItemDefinitionGroup, so that child ClCompile/Link/Lib/etc nodes will have their properties evaluated and skip calls to `Set-ProjectItemProperty` if the conditions are false.